Runtime Fix

+++ code/modules/admin/verbs/clear_toxins.dm
@@ -15,7 +15,12 @@
 	environment.temperature = 293.15
 	environment.update_values()
 	var/turf/simulated/location = get_turf(usr)
+	if (istype(location, /turf/space))
+		return
+
 	if (location.zone)
 		for (var/turf/T in location.zone.contents)
 			for (var/obj/fire/F in T.contents)
 				qdel(F)
+
+		log_and_message_admins("cleared air and fire in area [get_area(usr)].")
